Created simple versions of client and server

+++ SteTech1/src/com/scs/stevetech1/netmessages/NewEntityData.java
@@ -9,9 +9,8 @@ import com.scs.stevetech1.components.IEntity;
 public class NewEntityData {
 
 	public int entityID;
-	public byte type; // Entity code
+	public int type; // Entity code
 	public HashMap<String, Object> data = new HashMap<>(); // todo - use codes instead of string
-	//public boolean clientShouldAddImmed = false; // e.g. for players bullets, create them immediately - -todo - remove this and add imed if playerid > 0
 	
 	public NewEntityData() {
 		
@@ -19,7 +18,7 @@ public class NewEntityData {
 
 	public NewEntityData(IEntity e) {
 		this.entityID = e.getID();
-		this.type = (byte) e.getType(); // todo - check > 127
+		this.type = e.getType();
 		this.data = e.getCreationData();
 	}
